fix(tags): Fixed Add Tag modal rerendering on spots in map extent updated
