+++ src/modules/maps/Basemaps.js
@@ -119,7 +119,7 @@ const Basemap = (props) => {
   // Update spots in extent and saved view (center and zoom)
   const onMapIdle = async (e) => {
     console.log('Event onMapIdle', e);
-    props.spotsInMapExtent();
+    props.updateSpotsInMapExtent();
     if (!props.imageBasemap && !props.stratSection && mapRef?.current) {
       const newCenter = await mapRef.current.getCenter();
       const newZoom = await mapRef.current.getZoom();
