+++ src/modules/tags/index.js
@@ -1,14 +1,16 @@
-export {default as Tags} from './Tags';
-export {default as TagAddRemoveSpots} from './AddRemoveTagSpots';
+export {default as AddTagsToSpotsShortcutModal} from './AddTagsToSpotsShortcutModal';
+export {default as FeatureTagsModal} from './FeatureTagsModal';
 export {default as TagAddRemoveFeatures} from './AddRemoveTagFeatures';
-export {default as TagsAtSpotList} from './TagsAtSpotList';
+export {default as TagAddRemoveSpots} from './AddRemoveTagSpots';
 export {default as TagDetailModal} from './TagDetailModal';
 export {default as TagDetailSidePanel} from './TagDetailSidePanel';
 export {default as TagDetail} from './TagDetail';
+export {default as TagsAtSpotList} from './TagsAtSpotList';
+export {default as TagsListItem} from './TagsListItem';
+export {default as TagsList} from './TagsList';
 export {default as TagsModal} from './TagsModal';
 export {default as TagsNotebookModal} from './TagsNotebookModal';
 export {default as TagsShortcutModal} from './TagsShortcutModal';
-export {default as AddTagsToSpotsShortcutModal} from './AddTagsToSpotsShortcutModal';
+export {default as Tags} from './Tags';
 export {default as tagsStyles} from './tags.styles';
 export {default as useTagsHook} from './useTags';
-
